pyproject.toml
[tool.poetry]
name = "shoestore_drf"
version = "0.1.0"
description = ""
authors = ["Jaspal Singh <jsinghw94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
